Correct Oracle Instant Client build.
1) Oracle Instant Client package names changed after v10 releases to
include the major version number. Introduced variable for package name
that's changed depending on version so that package is correctly found.

2) Instant Client stores libraries under /usr/lib for both 32bit and
64bit builds. %{_libdir} caused incorrect path to be passed to configure
script. Corrected so that /usr/lib is used for 32bit and 64bit.

3) configure script option "--with-oci8" was incorrectly being passed
the Instant Client version. This is only an option for
"--with-pdo-oci8". Corrected so that Instant Client version is not
supplied as last argument.

4) Comment on line 630 should have apostrophe escaped. Corrected.